requirements.txt
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
-e git+ssh://git@github.com/gcantieni/gtunes.git@2e1980d5278a64b94195cd33677f7e4abb9c6600#egg=gtunes
idna==3.10
Levenshtein==0.26.1
python-dotenv==1.0.1
RapidFuzz==3.12.1
redis==5.2.1
requests==2.32.3
soupsieve==2.6
spotipy==2.25.0
urllib3==2.3.0
